[Design] : ADBackground DoodleLine 생성
Git : #6
